test complete

+++ src/main/java/com/auth/opinionscope/repository/UserRepository.java
@@ -3,10 +3,11 @@ package com.auth.opinionscope.repository;
 import com.auth.opinionscope.model.Questions;
 import com.auth.opinionscope.model.auth.UserData;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<UserData, Integer> {
+public interface UserRepository extends JpaRepository<UserData, Long> {
 //    List<User> findAll();
     Optional<UserData> findByEmail(String Email);
     Optional<UserData> findByUserId(Long UserId);
